Refactor: Hide 폴더 구조 변경
Related to:
