+++ src/main/java/com/service/dida/domain/hide/nft_hide/NftHide.java
@@ -1,4 +1,4 @@
-package com.service.dida.domain.hide;
+package com.service.dida.domain.hide.nft_hide;
 
 import com.service.dida.domain.member.entity.Member;
 import com.service.dida.domain.nft.Nft;
