+++ src/main/java/com/service/dida/domain/hide/nft_hide/service/UpdateNftHideService.java
@@ -1,8 +1,8 @@
-package com.service.dida.domain.hide.service;
+package com.service.dida.domain.hide.nft_hide.service;
 
-import com.service.dida.domain.hide.NftHide;
-import com.service.dida.domain.hide.repository.NftHideRepository;
-import com.service.dida.domain.hide.usecase.UpdateNftHideUseCase;
+import com.service.dida.domain.hide.nft_hide.NftHide;
+import com.service.dida.domain.hide.nft_hide.repository.NftHideRepository;
+import com.service.dida.domain.hide.nft_hide.usecase.UpdateNftHideUseCase;
 import com.service.dida.domain.member.entity.Member;
 import com.service.dida.domain.nft.Nft;
 import com.service.dida.domain.nft.repository.NftRepository;
