+++ src/main/java/com/service/dida/domain/hide/nft_hide/usecase/RegisterNftHideUseCase.java
@@ -1,4 +1,4 @@
-package com.service.dida.domain.hide.usecase;
+package com.service.dida.domain.hide.nft_hide.usecase;
 
 import com.service.dida.domain.member.entity.Member;
